[namespace] Moved library namespace to proper place
