+++ biz.aQute.bndlib/src/aQute/bnd/service/library/package-info.java
@@ -1,0 +1,2 @@
+@org.osgi.annotation.versioning.Version("1.0.0")
+package aQute.bnd.service.library;
